Enable image filters (doh). Add a detailed JPEG filter. (I got carried away...) This strips EXIF and comments, and is pretty thorough (although it doesn't parse *everything* e.g. the scan).
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@11206 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/io/CountedInputStream.java
@@ -29,7 +29,7 @@ public class CountedInputStream extends FilterInputStream {
         return ret;
     }
     
-    public long skip(int n) throws IOException {
+    public long skip(long n) throws IOException {
     	long l = in.skip(n);
     	if(l > 0) count += l;
     	return l;
